Warn that the scripts no longer work as expected
We lack time to spend on determining the best way to retrieve the
LE chain and import it into IPA automatically. The current hardcoding
of the chain has gone from bad to worse such that multiple people
have broken systems.

This is no longer used by the demo team so warn new users against using
it.

Signed-off-by: Rob Crittenden <rcritten@redhat.com>